Moved ScriptureLink styled component to the ScriptureList component where it is used.
