+++ packages/apolloschurchapp/src/ui/Scripture/ScriptureList.js
@@ -1,11 +1,30 @@
-const ScriptureList = ({ scripture, commas = true }) => {
+import React from 'react';
+import { H4 } from 'apolloschurchapp/src/ui/typography';
+import styled from 'apolloschurchapp/src/ui/styled';
+
+const ScriptureLink = styled(({ theme }) => ({
+  color: theme.colors.primary,
+  textAlign: 'center',
+  textDecorationLine: 'underline',
+  paddingBottom: theme.sizing.baseUnit * 1.5,
+}))(H4);
+
+const ScriptureList = ({ scripture, jumpTo, commas = true }) => {
   if (!scripture && commas) return '';
   if (!scripture) return [];
 
-  const combo = scripture.map(({ reference }) => `${reference}`);
+  let combo = scripture.map(({ reference }) => `${reference}`);
+  console.log('combo = ', combo);
 
-  if (commas) return combo.join(', ');
-  return combo;
+  if (commas) {
+    combo = combo.join(', ');
+  }
+
+  return (
+    <ScriptureLink padded onPress={() => jumpTo('scripture')}>
+      {combo}
+    </ScriptureLink>
+  );
 };
 
 export default ScriptureList;
